refactor #19 - 패키지 이름 변경
