+++ src/main/java/com/depromeet/todo/application/security/MockTokenService.java
@@ -1,4 +1,4 @@
-package com.depromeet.todo.application;
+package com.depromeet.todo.application.security;
 
 import org.springframework.stereotype.Service;
