+++ src/main/java/com/depromeet/todo/application/security/TokenService.java
@@ -1,4 +1,4 @@
-package com.depromeet.todo.application;
+package com.depromeet.todo.application.security;
 
 public interface TokenService<T> {
     String generate(T memberId);
